Download notified versions of files and bundles, not latest
For bundles this may be odd at best but for files this bad. A bundle always refers to a specific version of a file. Not downloading the specific version of a file that a bundle refers to, will produce incorrect results.

Also elininate chainmap and replace multiprocessing.dummy.ThreadPool with concurrent.futures.ThreadPoolExecutor.