Improve document merging (fixes #148)
The BaseIndexer.merge() method used to favor the current bundle, now it favors the indexed bundle. This is the root cause of #148 code as laid out by @caaespin. The method was irksome in other ways, too. For one it is now idempotent (it was cumulative before) and more general (it doesn't assume that the first bundle is the indexed one. Lastly it is now pure and does not modify its inputs in place.